+++ lib/euclidean/euclidean.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-unsigned int euclid(int n, int k, int o);
+uint16_t euclid(int n, int k, int o);
 uint16_t rightRotate(int shift, uint16_t value, uint8_t pattern_length);
 int findlength(unsigned int bnry);
 unsigned int ConcatBin(unsigned int bina, unsigned int binb);
